requirements.txt
# Local Autonomous Deep Research Agent (Lancet)
# Python 3.12+ Required

# === Core Dependencies ===
pydantic>=2.5.0
pydantic-settings>=2.1.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
aiohttp>=3.9.0
aiofiles>=23.2.1

# === MCP Server ===
mcp>=1.0.0

# === HTTP Client (Browser Impersonation) ===
curl-cffi>=0.6.0

# === Browser Automation ===
playwright>=1.40.0
# undetected-chromedriver for fallback (optional)
# undetected-chromedriver>=3.5.4

# === Content Extraction ===
trafilatura>=1.6.0
readability-lxml>=0.8.1
justext>=3.0.0
boilerpy3>=1.0.6

# === PDF Processing ===
pymupdf>=1.23.0
# camelot-py[cv]>=0.11.0  # Table extraction (optional)
# tabula-py>=2.8.0  # Table extraction (optional)

# === Japanese NLP ===
sudachipy>=0.6.7
sudachidict-core>=20231110

# === Text Processing & Search ===
rank-bm25>=0.2.2
tldextract>=5.1.0

# === Embeddings & Reranking (ONNX) ===
onnxruntime-gpu>=1.16.0  # GPU version, falls back to CPU
# onnxruntime>=1.16.0  # CPU-only alternative
transformers>=4.36.0
tokenizers>=0.15.0
sentence-transformers>=2.2.2
torch>=2.1.0  # For model loading

# === NLI (Natural Language Inference) ===
# tiny-deberta models loaded via transformers/onnx

# === Graph & Similarity ===
networkx>=3.2.0
datasketch>=1.6.0  # MinHash
# simhash>=2.1.0  # SimHash (optional)

# === Archive & WARC ===
warcio>=1.7.4

# === Tor Network ===
stem>=1.8.2
pysocks>=1.7.1

# === Database ===
aiosqlite>=0.19.0

# === Notification ===
# Windows toast via PowerShell (no pip dependency)
# Linux: libnotify (system package)

# === Logging & Monitoring ===
structlog>=23.2.0

# === Testing ===
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# === Development ===
black>=23.11.0
ruff>=0.1.6
mypy>=1.7.0

